XT-115 configuration for demo
Allows launch Configure request
Dialogs refactoring

XT-115
