+++ samples/hello-world/aspnetcore/ClientApp/src/app/Actions.tsx
@@ -4,10 +4,11 @@ import { observer } from 'mobx-react-lite';
 
 import { Grid, Typography, createStyles, makeStyles, Theme, Button } from '@material-ui/core';
 
-import AuthenticationAction from './AuthenticationAction';
-import NotifyAction from './NotifyAction';
-import DecorateAction from './DecorateAction';
-import TokenAction from './TokenAction';
+import AuthenticationAction from './dialogs/AuthenticationAction';
+import NotifyAction from './dialogs/NotifyAction';
+import DecorateAction from './dialogs/DecorateAction';
+import ConfigureAction from './dialogs/ConfigureAction';
+import TokenAction from './dialogs/TokenAction';
 import JsonView from './JsonView';
 import { EventStoreContext } from '../stores/EventStore';
 
@@ -45,6 +46,7 @@ const Actions: React.FC = observer(() => {
   
   const [notificationDialog, setNotificationDialog] = useState<boolean>(false);
   const [decorationDialog, setDecorationDialog] = useState<boolean>(false);
+  const [configurationDialog, setConfigurationDialog] = useState<boolean>(false);
   const [authorizationDialog, setAuthenticationDialog] = useState<boolean>(false);
   const [tokenDialog, setTokenDialog] = useState<boolean>(false);
 
@@ -80,6 +82,18 @@ const Actions: React.FC = observer(() => {
         </Grid>
         <Grid item={true} className={classes.item}>
           <Button
+              autoCapitalize="false"
+              variant="outlined"
+              type="submit"
+              className={classes.actionButton}
+              color="primary"
+              onClick={e => setConfigurationDialog(true)}
+            >
+              sdk.configure()
+          </Button>
+        </Grid>
+        <Grid item={true} className={classes.item}>
+          <Button
               variant="outlined"
               type="submit"
               className={classes.actionButton}
@@ -103,6 +117,7 @@ const Actions: React.FC = observer(() => {
       </Grid>
       <NotifyAction open={notificationDialog} onClose={() => setNotificationDialog(false)} />
       <DecorateAction open={decorationDialog} onClose={() => setDecorationDialog(false)} />
+      <ConfigureAction open={configurationDialog} onClose={() => setConfigurationDialog(false)} />
       <AuthenticationAction open={authorizationDialog} onClose={() => setAuthenticationDialog(false)} />
       <TokenAction open={tokenDialog} onClose={() => setTokenDialog(false)} />
       {eventStore.json && <JsonView json={eventStore.json} onClose={() => eventStore.setJson(null)} />}
